playbooks/roles/ilmnuri/files/requirements.txt
awscli
botocore
Flask
python-memcached
requests
uwsgi
setuptools
python-geoip-geolite2
